refactor: applicationOrder 자료형을 long 으로 수정

+++ src/main/java/com/example/solidconnection/admin/dto/MentorApplicationHistoryResponse.java
@@ -8,7 +8,7 @@ public record MentorApplicationHistoryResponse(
         MentorApplicationStatus mentorApplicationStatus,
         String rejectedReason,
         ZonedDateTime createdAt,
-        int applicationOrder
+        long applicationOrder
 ) {
 
 }
